Introduce use of bundler; simplify tests
Previously, we created a new TCP connection to export every row.
In a large application, this could easily mean hundreds of TCP
connections set up and torn down every reporting period.

This change introduces a bundling mechanism as used in the
Stackdriver exporter to try to reuse the same connection for
multiple exported time series. It still opens and closes a connection
per report call, but this is much less than for every data point.

Also simplified the tests by using an in-memory buffer instead of
a real TCP connection.